+++ openml/extensions/sklearn/__init__.py
@@ -1,4 +1,4 @@
-from .sklearn_extension import SklearnExtension
+from .extension import SklearnExtension
 
 
 __all__ = ['SklearnExtension']
